pyproject.toml
# SPDX-FileCopyrightText: 2022 Bart Ribbers <bribbers@disroot.org>
# SPDX-License-Identifier: CC0-1.0

[tool.poetry]
name = "aportsknife"
version = "0.0.1"
description = "A 'swiss knife' tool for Alpine Linux aports bulk operations"
authors = [
    "Bart Ribbers <bribbers@disroot.org>",
    ]
packages = [
    { include = "aportsknife" }
]

[tool.poetry.scripts]
aportsknife = "aportsknife:main"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pyxdg = "0.28"
GitPython = "3.1.27"
PyYAML = "6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-print = "^5.0.0"
pep8-naming = "^0.12.1"
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.961"

[build-system]
requires = ["poetry-core>=1.0.5"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
skip = [".venv"]
profile = "black"

[tool.black]
line-length = 120